AI - Add a new Maps tab, along with new data models and sql tables for maps. Maps will have an Attributes functionality similar to Builds, but all maps will have the same set of attributes. The UI should be similar to Builds, but Maps cannot be nested, unlike Builds. The Map selection on the left should have a textbox for filtering by name at the top, with the list of Maps below it. When a map is selected, show the Attributes on the right. Adding a new Attribute to a map defines the attribute for all Maps, but will not have a default value (should be null for all maps). Next to the filter textbox, there should be an add filter button, similar to the one on the DataPage, which will let you filter by any defined Map Attribute.
Games should reference a Map object instead of using a string with just the map name. If no Map exists for the map name, create a new Map entry for that name.

Note - ran this on Opus and it took 60% of my 5-hour token limit. I get that this was a somewhat complex request, but that still seems high. Switching back to Sonnet

+++ StatCraft/StatCraft/ViewModels/Windows/DataComponents/NumericRangeFilterSlotViewModel.cs
@@ -2,18 +2,22 @@ using CommunityToolkit.Mvvm.ComponentModel;
 
 namespace StatCraft.ViewModels
 {
-    // An extra filter whose criteria is a numeric [Min, Max] range (opponent MMR).
+    // An extra filter whose criteria is a numeric [Min, Max] range — opponent MMR on the Data tab, and
+    // any Numeric or Percent map attribute on the Maps tab.
+    //
+    // decimal rather than long because map attribute values are decimal; MMR is integral, so the Data
+    // tab narrows back to long when it builds its criteria.
     public sealed partial class NumericRangeFilterSlotViewModel : FilterSlotViewModel
     {
-        [ObservableProperty] private long? _min;
-        [ObservableProperty] private long? _max;
+        [ObservableProperty] private decimal? _min;
+        [ObservableProperty] private decimal? _max;
 
         internal NumericRangeFilterSlotViewModel(string title) : base(title)
         {
         }
 
-        partial void OnMinChanged(long? value) => RaiseChanged();
-        partial void OnMaxChanged(long? value) => RaiseChanged();
+        partial void OnMinChanged(decimal? value) => RaiseChanged();
+        partial void OnMaxChanged(decimal? value) => RaiseChanged();
 
         public override void Clear()
         {
